+++ test/HCurve.test.js
@@ -1103,7 +1103,7 @@ contract('Curve', function([_, user]) {
           answer.mul(new BN('999')).div(new BN('1000'))
         );
         expect(await this.poolToken.balanceOf.call(user)).to.be.bignumber.lte(
-          answer
+          answer.mul(new BN('1010')).div(new BN('1000'))
         );
 
         profileGas(receipt);